testing partial classes for plugin-API-extension grouping on rmmz data;
