+++ src/Dashboard/Boards/SkillsBoard.cs
@@ -1,5 +1,5 @@
-using Rmmz.Data.Models.db.implementations;
 using Rmmz.Data.Models.JABS;
+using Rmmz.Data.Models.JABS.implementations;
 
 namespace Dashboard.Boards;
