+++ src/Dashboard/Services/JsonLoaderService.cs
@@ -1,5 +1,5 @@
 ﻿using Newtonsoft.Json;
-using Rmmz.Data.Models.db.implementations;
+using Rmmz.Data.Models.JABS.implementations;
 
 namespace Dashboard.Services;
