+++ src/Dashboard/Services/JsonSavingService.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
-using Rmmz.Data.Models.db.implementations;
+using Rmmz.Data.Models.JABS.implementations;
 
 namespace Dashboard.Services;
